08-add_domain_events_to_legacy/2-add_event_bus/1-without_events/src/contexts/shop/users/application/registrar/UserRegistrar.ts
import { WelcomeEmailSender } from "../../../../retention/email/application/send_welcome_email/WelcomeEmailSender";
import { User } from "../../domain/User";
import { UserRepository } from "../../domain/UserRepository";

export class UserRegistrar {
	constructor(
		private readonly repository: UserRepository,
		private readonly emailSender: WelcomeEmailSender,
	) {}

	async registrar(id: string, name: string, email: string, profilePicture: string): Promise<void> {
		const user = User.create(id, name, email, profilePicture);

		await this.repository.save(user);
		await this.emailSender.send(id, name, email);
	}
}
